# docs/api/reference-data-service-openapi.yaml
openapi: 3.0.3
info:
  title: Reference Data Service API
  description: |
    Authoritative read API for Security, Account, and Book reference masters consumed by domain services
    (Trade Capture, Position, Lifecycle) via the shared Reference Data Client SDK.
  version: 1.0.0

servers:
  - url: /api/v1
    description: Version 1 (relative to service root)

tags:
  - name: Securities
  - name: Accounts
  - name: Books

paths:
  /securities/{securityId}:
    get:
      tags: [Securities]
      operationId: getSecurityById
      parameters:
        - $ref: '#/components/parameters/securityId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Security'
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /securities/by-ric:
    get:
      tags: [Securities]
      operationId: getSecurityByRic
      parameters:
        - name: ric
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Security'
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /securities/batch:
    post:
      tags: [Securities]
      operationId: batchGetSecurities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchSecurityRequest'
      responses:
        '200':
          description: OK — order of results matches request ids where found; omit or null for missing (implementation returns found only with id on each item).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchSecurityResponse'

  /accounts/{accountId}:
    get:
      tags: [Accounts]
      operationId: getAccountById
      parameters:
        - $ref: '#/components/parameters/accountId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /accounts/batch:
    post:
      tags: [Accounts]
      operationId: batchGetAccounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchAccountRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchAccountResponse'

  /books/{bookId}:
    get:
      tags: [Books]
      operationId: getBookById
      parameters:
        - $ref: '#/components/parameters/bookId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /books/by-code:
    get:
      tags: [Books]
      operationId: getBookByCode
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /books/batch:
    post:
      tags: [Books]
      operationId: batchGetBooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchBookRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchBookResponse'

components:
  parameters:
    securityId:
      name: securityId
      in: path
      required: true
      schema:
        type: string
    accountId:
      name: accountId
      in: path
      required: true
      schema:
        type: string
    bookId:
      name: bookId
      in: path
      required: true
      schema:
        type: string

  schemas:
    Problem:
      type: object
      description: RFC 7807 Problem Details
      properties:
        type:
          type: string
          format: uri
          example: https://pb.synth/problems/not-found
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          description: Stable machine-readable code e.g. SECURITY_NOT_FOUND
          example: SECURITY_NOT_FOUND

    Security:
      type: object
      required: [securityId, version]
      properties:
        securityId:
          type: string
        ric:
          type: string
        isin:
          type: string
        currency:
          type: string
        assetType:
          type: string
        description:
          type: string
        version:
          type: integer
          format: int64
          description: Optimistic versioning / cache invalidation

    BatchSecurityRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 200

    BatchSecurityResponse:
      type: object
      required: [securities]
      properties:
        securities:
          type: array
          items:
            $ref: '#/components/schemas/Security'

    Account:
      type: object
      required: [accountId, version]
      properties:
        accountId:
          type: string
        name:
          type: string
        classification:
          type: string
        creditTier:
          type: string
        stpEligible:
          type: boolean
        version:
          type: integer
          format: int64

    BatchAccountRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 200

    BatchAccountResponse:
      type: object
      required: [accounts]
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/Account'

    Book:
      type: object
      required: [bookId, version]
      properties:
        bookId:
          type: string
        code:
          type: string
          description: Business book code used in GCAM / allocation messages
        entityName:
          type: string
        desk:
          type: string
        normalisedKey:
          type: string
          description: Key used in partition / routing after enrichment
        version:
          type: integer
          format: int64

    BatchBookRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 200

    BatchBookResponse:
      type: object
      required: [books]
      properties:
        books:
          type: array
          items:
            $ref: '#/components/schemas/Book'